moved check_cuts as a method of spectral_network
